Divide index into action js file
